+++ spec/support/fixtures/with_registry.rb
@@ -94,7 +94,6 @@ module Foo
       after "externally-extended", ::ExternallyExtended::Callback
       option "externally-extended", :skip_tests, type: :flag, default: false,
         desc: "Skip test generation"
-      argument "externally-extended", :suite, required: false, desc: "Test suite"
 
       before("externally-extended") do |args|
         puts "before block: skip_tests: #{args.fetch(:skip_tests)}"