+++ spec/support/fixtures/with_zero_arity_block.rb
@@ -97,7 +97,6 @@ WithZeroArityBlock = Dry.CLI do
   after "externally-extended", ExternallyExtended::Callback
   option "externally-extended", :skip_tests, type: :flag, default: false,
     desc: "Skip test generation"
-  argument "externally-extended", :suite, required: false, desc: "Test suite"
 
   before("externally-extended") do |args|
     puts "before block: skip_tests: #{args.fetch(:skip_tests)}"